Simplify trigger logic
This change makes it so all the 'calculation' of what commits to build are done
in git land by passing appropriate arguments to rev-list.  This means things
are more correct and more performant since only a single revwalk is needed.

The logic does still need to be synchronous to ensure refs don't change during
the revwalk, but the synchronous hook PostReceiveHook does not get run when
pull requests are merged (!!!) so we have to catch that edge case elsewhere.
We do that by also triggering builds when we see the PullRequestMergedEvent.
See PullRequestListener for that code.