Remove the display of default value
We can't show the default value because this value is set on a per-jenkins
server basis.  Since a user can modify which server they're using, this
will require some kind of dynamic modification of the page.

In the future, perhaps we can load all the default timeouts for each server
(somehow) in the soy template and then react to dropdown changes.