Added the ability for error messages in repo config
-In my testing of build timeout ranges, the page would just error out
-The Jenkins Settings had error messages so I used its method which
  checks for errors and redirects with the error message
-This will be better so people don't WTF when they can't see the exception
  from the atlassian log

Be more restrictive about what errors we expose to the user in Repo config

We should only expose errors that are relevant to the user herself.  For
example, show the NumberFormatException for parseInt(String) if they enter
a bad integer value.  If something more catastrophic happens, they shouldn't
be getting weird low-level DB errors that they didn't cause.  The exception
will still get shown in the stash log.